Address nits
